Separate merkle module
