+++ src/bin/inx-chronicle/api/stardust/poi/mod.rs
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod error;
-mod merkle_hasher;
-mod merkle_proof;
+mod merkle;
 mod responses;
 mod routes;
